+++ src/ray.rs
@@ -18,7 +18,7 @@ impl Ray {
         self.direction
     }
 
-    pub(crate) fn at(&self, t: f32) -> Point3 {
+    pub(crate) fn at(&self, t: f64) -> Point3 {
         self.origin + t * self.direction
     }
 }
